Move pages from OMIS edit/client directory to modules
