--- src/apps/omis/apps/edit/client/state.js
@@ -1,2 +1,0 @@
-export const TASK_SAVE_ORDER_ASSIGNEES = 'TASK_SAVE_ORDER_ASSIGNEES'
-export const TASK_SAVE_ORDER_SUBSCRIBERS = 'TASK_SAVE_ORDER_SUBSCRIBERS'
