+++ src/client/modules/Omis/state.js
@@ -10,6 +10,8 @@ export const TASK_CANCEL_ORDER = 'TASK_CANCEL_ORDER'
 export const TASK_EDIT_ORDER_ASSIGNEE_TIME = 'TASK_EDIT_ORDER_ASSIGNEE_TIME'
 export const TASK_COMPLETE_ORDER = 'TASK_COMPLETE_ORDER'
 export const TASK_EDIT_ORDER_CONTACT = 'TASK_EDIT_ORDER_CONTACT'
+export const TASK_SAVE_ORDER_ASSIGNEES = 'TASK_SAVE_ORDER_ASSIGNEES'
+export const TASK_SAVE_ORDER_SUBSCRIBERS = 'TASK_SAVE_ORDER_SUBSCRIBERS'
 
 export const EDIT_CONTACT_ID = 'editOmisContactId'
