+++ src/client/modules/Omis/tasks.js
@@ -1,4 +1,8 @@
 import { apiProxyAxios } from '../../components/Task/utils'
+import {
+  transformAdvisersForAPI,
+  transformSubscribersForAPI,
+} from './transformers'
 
 export const updateOrder = (values) => {
   const useToken = values.csrfToken ? `?_csrf=${values.csrfToken}` : ''
@@ -19,3 +23,17 @@ export const completeOrder = (values) =>
     `v3/omis/order/${values.id}/assignee`,
     values.assignees
   ) && apiProxyAxios.post(`v3/omis/order/${values.id}/complete`)
+
+export function saveOrderAssignees({ values, id, canRemoveAssignees }) {
+  const url = canRemoveAssignees
+    ? `/v3/omis/order/${id}/assignee?force-delete=1`
+    : `/v3/omis/order/${id}/assignee`
+  return apiProxyAxios.patch(url, transformAdvisersForAPI(values))
+}
+
+export function saveOrderSubscribers({ values, id, canRemoveSubscribers }) {
+  const url = canRemoveSubscribers
+    ? `/v3/omis/order/${id}/subscriber-list?force-delete=1`
+    : `/v3/omis/order/${id}/subscriber-list`
+  return apiProxyAxios.put(url, transformSubscribersForAPI(values))
+}
